requirement.txt
# requirements.txt
flask==3.0.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9  # Conector oficial para PostgreSQL